requirements.txt
requests
prometheus-client
python-dotenv